Exercise 2-7: keywords
Keywords are predefined or reserved words that have special meanings to
the compiler. These are part of the syntax and cannot be used as
identifiers in the program.

From the given examples:
a) 'for' is a keyword.
b) 'If' is not a keyword. c is case-sensitive, the 'if' keyword is
lower-case.
c) 'main' is not a keyword. This is the name of the functions called at
program startup.
d) 'printf' is not a keyword. 'printf' is the name of a function defined
in C's standard library.
e) 'while' is a keyword.