Exercise 8-1: sizeof(a) / sizeof(a[0])
Using sizeof(a) / sizeof(t), where t is the type of a's elements,
requires the programmer to update the type used in the expression if the
type of the array is changed. Updating the type in the expression could
easily be missed which would result in and incorrect value being
returned. Using the expression size(a) / size(a[0]) avoids this
potential pitfall because it will continue to return the intended result
when the array's type is updated and is therefore considered a superior
technique.