Exercise 8-2: digit character as array subscript
A digit in character form can be used as an array subscript by
subtracting the integer value of the character '0' from the digit
character being used. This will work if the character set being used has
sequential integer values for the character 0 - 9.

For example, to access the 5th element of an array:

char c = '5';
array[c - '0'];